🔨 refactor : 클라이언트 훅을 하위 컴포넌트로 분리하여 페이지에서 use client 분리
